#72 -  moved method to retrieve nillable properties to DataSampler
